# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Star Wars REST API with Swagger",
    "version": "0.1.0",
    "description": "Welcome to the swapi, the Star Wars API! This documentation should help you familiarise yourself with the resources available and how to consume them with HTTP requests.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Corentin Goujon-Cholet",
      "email": "corentingcholet@gmail.com"
    }
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Starship": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "starship_class": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "cost_in_credits": {
            "type": "string"
          },
          "length": {
            "type": "string"
          },
          "crew": {
            "type": "string"
          },
          "passengers": {
            "type": "string"
          },
          "max_atmosphering_speed": {
            "type": "string"
          },
          "hyperdrive_rating": {
            "type": "string"
          },
          "MGLT": {
            "type": "string"
          },
          "cargo_capacity": {
            "type": "string"
          },
          "consumables": {
            "type": "string"
          },
          "films": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pilots": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "url": {
            "type": "string"
          }
        }
      },
      "Vehicle": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "vehicle_class": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "cost_in_credits": {
            "type": "string"
          },
          "crew": {
            "type": "string"
          },
          "passengers": {
            "type": "string"
          },
          "max_atmosphering_speed": {
            "type": "string"
          },
          "cargo_capacity": {
            "type": "string"
          },
          "consumables": {
            "type": "string"
          },
          "films": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pilots": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "url": {
            "type": "string"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/api/starship": {
      "post": {
        "tags": [
          "Starship"
        ],
        "description": "Get all starships",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Starship"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Starship"
        ],
        "description": "Get all starship",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Starship"
        ],
        "description": "Delete all Starships",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/starship/{id}": {
      "get": {
        "tags": [
          "Starship"
        ],
        "description": "Get starship by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The starship id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "put": {
        "tags": [
          "Starship"
        ],
        "description": "Modify starship by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The starship id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Starship"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Starship"
        ],
        "description": "Delete starship by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The starship id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/vehicle": {
      "post": {
        "tags": [
          "Vehicle"
        ],
        "description": "Get all vehicles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Vehicle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Vehicle"
        ],
        "description": "Get all vehicle",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Vehicle"
        ],
        "description": "Delete all Vehicles",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/vehicle/{id}": {
      "get": {
        "tags": [
          "Vehicle"
        ],
        "description": "Get vehicle by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The vehicle id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "put": {
        "tags": [
          "Vehicle"
        ],
        "description": "Modify vehicle by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The vehicle id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Vehicle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Vehicle"
        ],
        "description": "Delete vehicle by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The vehicle id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    }
  }
}